[REFACTOR] update folder product test
